Load solution

+++ DataPoint.cpp
@@ -11,7 +11,7 @@ DataPoint::DataPoint(unsigned int customer,unsigned int x, unsigned int y, unsig
         :customer_number(customer), x_coordinate(x), y_coordinate(y), demand(demand),
         ready_time(ready),due_date(due), service(service) {}
 
-float DataPoint::get_distance(const DataPoint &dataPoint) const {
+double DataPoint::get_distance(const DataPoint &dataPoint) const {
     double dx = static_cast<double>(dataPoint.x_coordinate) - static_cast<double>(x_coordinate);
     double dy = static_cast<double>(dataPoint.y_coordinate) - static_cast<double>(y_coordinate);
 
@@ -35,14 +35,14 @@ unsigned int DataPoint::getDemand() const {
     return demand;
 }
 
-float DataPoint::load_time(float previous_load_time, const DataPoint & previous_vertex) const {
+double DataPoint::load_time(double previous_load_time, const DataPoint & previous_vertex) const {
     /// b_i = max{ b_i-1 + d_i-1 + c_i-1,i , e_i  }
     /// b_1-1 czas zakończenia rozładunku w poprzednim wierzchołku.
     ///d_i-1 czas rozładunku w poprzednim wierzchołku.
     ///c_i-1,i czas przejazdu między wierzchołkami i-1 i i
     ///e_i  najwcześniejszy dopuszczalny czas rozładunku w wierzchołku i
 
-    let next_available_time = static_cast<float>(ready_time);
+    let next_available_time = static_cast<double>(ready_time);
 
 
     let finish_time = previous_load_time;
@@ -50,7 +50,7 @@ float DataPoint::load_time(float previous_load_time, const DataPoint & previous_
 
 
     return std::max(finish_time  + distance  ,
-        next_available_time)+ static_cast<float>(service);
+        next_available_time)+ static_cast<double >(service);
 }
 
 unsigned int DataPoint::getCustomerNumber() const {
